requirements.txt

lxml
python-docx
Pillow
sv-ttk 
python-json-logger
pyyaml